requirements.txt
# Database
mysql-connector-python>=8.0.26

# Web Framework
Flask>=2.0.1
Flask-Login>=0.5.0
Flask-WTF>=1.0.0

# Security
bcrypt>=3.2.0
python-dotenv>=0.19.0

# Development
pytest>=6.2.5
pylint>=2.11.1
